renames ControllerSettings
